fix: separating api server tests, and splitting the operator tests
closes: #37853
