+++ .github/workflows/operator-ci.yml
@@ -56,6 +56,9 @@ jobs:
     name: Test local
     runs-on: ubuntu-latest
     needs: [build]
+    strategy:
+      matrix:
+        suite: [slow, fast]
     steps:
       - uses: actions/checkout@v4
 
@@ -90,16 +93,23 @@ jobs:
 
       - name: Test operator running locally
         run: |
+          declare -A PARAMS
+          PARAMS["slow"]="-Dkc.quarkus.tests.groups=slow"
+          PARAMS["fast"]='-Dkc.quarkus.tests.groups=!slow'
+          
           ./mvnw install -Poperator -pl :keycloak-operator -am \
               -Dquarkus.kubernetes.image-pull-policy=IfNotPresent \
               -Dkc.operator.keycloak.image=keycloak:${{ env.version_local }} \
               -Dtest.operator.custom.image=custom-keycloak:${{ env.version_local }} \
-              -Dkc.operator.keycloak.image-pull-policy=Never
+              -Dkc.operator.keycloak.image-pull-policy=Never ${PARAMS["${{ matrix.suite }}"]}
 
   test-remote:
     name: Test remote
     runs-on: ubuntu-latest
     needs: [build]
+    strategy:
+          matrix:
+            suite: [slow, fast]
     steps:
       - uses: actions/checkout@v4
 
@@ -134,6 +144,10 @@ jobs:
 
       - name: Test operator running in cluster
         run: |
+          declare -A PARAMS
+          PARAMS["slow"]="-Dkc.quarkus.tests.groups=slow"
+          PARAMS["fast"]='-Dkc.quarkus.tests.groups=!slow'
+          
           eval $(minikube -p minikube docker-env)
           ./mvnw install -Poperator -pl :keycloak-operator -am \
               -Dquarkus.container-image.build=true \
@@ -141,7 +155,7 @@ jobs:
               -Dkc.operator.keycloak.image=keycloak:${{ env.version_remote }} \
               -Dquarkus.kubernetes.env.vars.kc-operator-keycloak-image-pull-policy=Never \
               -Dtest.operator.custom.image=custom-keycloak:${{ env.version_remote }} \
-              --no-transfer-progress -Dtest.operator.deployment=remote
+              --no-transfer-progress -Dtest.operator.deployment=remote ${PARAMS["${{ matrix.suite }}"]}
 
   test-olm:
     name: Test OLM installation
