+++ operator/src/test/java/org/keycloak/operator/testsuite/integration/UpgradeTest.java
@@ -27,6 +27,8 @@ import java.util.concurrent.TimeoutException;
 import io.fabric8.kubernetes.api.model.batch.v1.Job;
 import io.fabric8.kubernetes.api.model.batch.v1.JobStatus;
 import io.quarkus.test.junit.QuarkusTest;
+
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -50,6 +52,7 @@ import static org.keycloak.operator.testsuite.utils.CRAssert.eventuallyRecreateU
 import static org.keycloak.operator.testsuite.utils.CRAssert.eventuallyRollingUpgradeStatus;
 import static org.keycloak.operator.testsuite.utils.K8sUtils.deployKeycloak;
 
+@Tag(BaseOperatorTest.SLOW)
 @QuarkusTest
 public class UpgradeTest extends BaseOperatorTest {
